# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "DND Character Backend",
    "version": "1.0.0",
    "description": "Snapshot + stable-ID operations. Protocol v1. Default limits documented in backend/README.md. JSON syntax is valid YAML."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "me",
        "summary": "Authenticated identity",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/session": {
      "post": {
        "operationId": "sessionCreate",
        "summary": "Exchange Bearer token for HttpOnly SameSite=Strict session cookie",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "dnd_session token; HttpOnly; SameSite=Strict; Path=/api/v1; Max-Age=86400; Secure when configured"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "sessionDelete",
        "summary": "Clear browser cookie; does not revoke provisioning token",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "204": {
            "description": "Success, no response body"
          }
        }
      }
    },
    "/api/v1/characters": {
      "get": {
        "operationId": "charactersList",
        "summary": "Permission-filtered metadata only",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "characters": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CharacterSummary"
                      }
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 9007199254740991
                    },
                    "offset": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "characters",
                    "limit",
                    "offset"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ]
      },
      "post": {
        "operationId": "characterCreate",
        "summary": "Create/import native Character v1; new server UUID and revision 1",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                },
                "example": {
                  "id": "40000000-0000-4000-8000-000000000001",
                  "ownerId": "50000000-0000-4000-8000-000000000001",
                  "system": "dnd5e",
                  "schemaVersion": 1,
                  "revision": 1,
                  "document": {
                    "schemaVersion": 1,
                    "id": "10000000-0000-4000-8000-000000000001",
                    "revision": 1,
                    "name": "协议测试角色",
                    "player": "",
                    "edition": "2024",
                    "createdAt": "2026-10-01T00:00:00Z",
                    "updatedAt": "2026-10-01T00:00:00Z",
                    "abilities": {
                      "str": 10,
                      "dex": 10,
                      "con": 10,
                      "int": 10,
                      "wis": 10,
                      "cha": 10
                    },
                    "baseHp": 20,
                    "identity": {
                      "gender": "",
                      "alignment": "",
                      "age": "",
                      "description": ""
                    },
                    "selections": [],
                    "answers": {},
                    "reviewed": [],
                    "profile": {
                      "enabledSources": [],
                      "optional": {
                        "feats": true,
                        "multiclass": false,
                        "legacy": false
                      },
                      "exceptions": {}
                    },
                    "notes": "原创合成测试数据",
                    "runtime": {
                      "hp": 20,
                      "tempHp": 0,
                      "inspiration": 0,
                      "resources": {}
                    }
                  },
                  "createdAt": "2026-10-01T00:00:00Z",
                  "updatedAt": "2026-10-01T00:00:00Z"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              },
              "example": {
                "document": {
                  "schemaVersion": 1,
                  "id": "10000000-0000-4000-8000-000000000001",
                  "revision": 1,
                  "name": "协议测试角色",
                  "player": "",
                  "edition": "2024",
                  "createdAt": "2026-10-01T00:00:00Z",
                  "updatedAt": "2026-10-01T00:00:00Z",
                  "abilities": {
                    "str": 10,
                    "dex": 10,
                    "con": 10,
                    "int": 10,
                    "wis": 10,
                    "cha": 10
                  },
                  "baseHp": 20,
                  "identity": {
                    "gender": "",
                    "alignment": "",
                    "age": "",
                    "description": ""
                  },
                  "selections": [],
                  "answers": {},
                  "reviewed": [],
                  "profile": {
                    "enabledSources": [],
                    "optional": {
                      "feats": true,
                      "multiclass": false,
                      "legacy": false
                    },
                    "exceptions": {}
                  },
                  "notes": "原创合成测试数据",
                  "runtime": {
                    "hp": 20,
                    "tempHp": 0,
                    "inspiration": 0,
                    "resources": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/characters/{id}": {
      "get": {
        "operationId": "characterSnapshot",
        "summary": "Owner/editor/viewer snapshot",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                },
                "example": {
                  "id": "40000000-0000-4000-8000-000000000001",
                  "ownerId": "50000000-0000-4000-8000-000000000001",
                  "system": "dnd5e",
                  "schemaVersion": 1,
                  "revision": 1,
                  "document": {
                    "schemaVersion": 1,
                    "id": "10000000-0000-4000-8000-000000000001",
                    "revision": 1,
                    "name": "协议测试角色",
                    "player": "",
                    "edition": "2024",
                    "createdAt": "2026-10-01T00:00:00Z",
                    "updatedAt": "2026-10-01T00:00:00Z",
                    "abilities": {
                      "str": 10,
                      "dex": 10,
                      "con": 10,
                      "int": 10,
                      "wis": 10,
                      "cha": 10
                    },
                    "baseHp": 20,
                    "identity": {
                      "gender": "",
                      "alignment": "",
                      "age": "",
                      "description": ""
                    },
                    "selections": [],
                    "answers": {},
                    "reviewed": [],
                    "profile": {
                      "enabledSources": [],
                      "optional": {
                        "feats": true,
                        "multiclass": false,
                        "legacy": false
                      },
                      "exceptions": {}
                    },
                    "notes": "原创合成测试数据",
                    "runtime": {
                      "hp": 20,
                      "tempHp": 0,
                      "inspiration": 0,
                      "resources": {}
                    }
                  },
                  "createdAt": "2026-10-01T00:00:00Z",
                  "updatedAt": "2026-10-01T00:00:00Z"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/api/v1/characters/{id}/permissions": {
      "get": {
        "operationId": "permissionsList",
        "summary": "Owner only",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or resync required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permissions"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dnd_session"
      }
    },
    "schemas": {
...
      "CharacterSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "system": {
            "const": "dnd5e"
          },
          "schemaVersion": {
            "const": 1
          },
          "revision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "ownerId",
          "system",
          "schemaVersion",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "validation_error",
              "unauthorized",
              "forbidden",
              "not_found",
              "revision_conflict",
              "schema_mismatch",
              "resync_required",
              "duplicate_operation",
              "invalid_operation",
              "invalid_path",
              "payload_too_large",
              "rate_limited",
              "internal_error"
            ]
          },
          "message": {
            "type": "string"
          },
          "currentRevision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "baseRevision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Conflict"
            }
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "Identity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "additionalProperties": false
      },
...
      "CreateRequest": {
        "type": "object",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/Character"
          }
        },
        "required": [
          "document"
        ],
        "additionalProperties": false
      }
    }
  }
}